Rename proptypes -> types
