+++ cmp/input/index.ts
@@ -1,3 +1,3 @@
 export * from './HoistInputModel';
 export * from './HoistInputProps';
-export * from './proptypes';
+export * from './types';
