+++ cmp/input/types/desktop/index.ts
