+++ cmp/input/types/mobile/MobileTextInputProps.ts
